requirements.txt
mcp
yfinance==0.2.36
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
aiohttp>=3.8.0
requests>=2.28.0
textblob>=0.17.0
slack-bolt>=1.18.0
slack-sdk>=3.20.0
python-dotenv>=1.0.0
threading>=0.1.0
asyncio-mqtt>=0.11.0
